+++ src/routed/KarangosForm.js
@@ -1,103 +1,252 @@
-import axios from 'axios'
-import { useEffect, useState } from 'react'
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
+import { useState, useEffect } from 'react'
+import TextField from '@material-ui/core/TextField'
+import MenuItem from '@material-ui/core/MenuItem'
+import { makeStyles } from '@material-ui/core/styles'
+import { Button, FormControl, Toolbar } from '@material-ui/core'
+import FormControlLabel from '@material-ui/core/FormControlLabel'
 import Checkbox from '@material-ui/core/Checkbox'
-import IconButton from '@material-ui/core/IconButton';
-import DeleteIcon from '@material-ui/icons/Delete';
-import EditIcon from '@material-ui/icons/Edit';
-import Toolbar from '@material-ui/core/Toolbar'
-import Button from '@material-ui/core/Button'
-import AddBoxIcon from '@material-ui/icons/AddBox';
-import { useHistory } from 'react-router-dom'
-import ConfirmDialog from '../ui/ConfirmDialog'
+import InputMask from 'react-input-mask'
+import InputAdornment from '@material-ui/core/InputAdornment'
+import axios from 'axios'
 import Snackbar from '@material-ui/core/Snackbar';
 import MuiAlert from '@material-ui/lab/Alert';
-import KarangosList from './KarangosList';
+import { useHistory, useParams } from 'react-router-dom'
+import ConfirmDialog from '../ui/ConfirmDialog'
 
-const useStyles = makeStyles((theme) => ({
-  table: {
-    minWidth: 650,
-  },
-  tableRow: {
-    '& button': {       // Linha da tabela em estado "normal"
-      visibility: 'hidden'
-    },
-    '&:hover': {        // Linha da tabela com mouse sobreposto
-      backgroundColor: theme.palette.action.hover
-    },
-    '&:hover button': { // Botões na linha com mouse sobreposto
-      visibility: 'visible'
+
+const useStyles = makeStyles(() => ({
+  form: {
+    maxWidth: '90%',
+    margin: '0 auto',
+    display: 'flex',
+    justifyContent: 'space-around',
+    flexWrap: 'wrap',
+    '& .MuiFormControl-root': {
+      minWidth: '200px',
+      maxWidth: '500px',
+      marginBottom: '24px',
     }
   },
   toolbar: {
-    justifyContent: 'flex-end',
-    paddingRight: 0
+    margin: '36px',
+    width: '100%',
+    display: 'flex',
+    justifyContent: 'space-around'
+  },
+  checkBox: {
+    alignItems: 'center'
   }
-}));
+}))
 
-export default function KarangosList() {
-  const classes = useStyles();
+/* Classes de caracteres de entrada para a máscara do campo placa
+1) Três primeiras posições: qualquer letra, de A a Z (maiúsculo ou minúsculo) ~> [A-Za-z]
+2) Posições numéricas (1ª, a 3ª e a 4ª depois do traço) ~> [0-9]
+3) 2ª posição após o traço: pode receber dígitos ou letras de A a J (maiúsculas ou minúsculas) ~> [0-9A-Ja-j]
+*/
 
-  const history = useHistory()
+// Representando as classes de caracteres da máscara como um objeto
+const formatChars = {
+  'A': '[A-Za-z]',
+  '0': '[0-9]',
+  '#': '[0-9A-Ja-j]'
+}
 
-  // É importante inicializar esta variável de estado como um vetor vazio
-  const [karangos, setKarangos] = useState([])
-  const [dialogOpen, setDialogOpen] = useState(false)
-  const [deletable, setDeletable] = useState() // Cód. do registro a ser excluído
+// Finalmente, a máscara de entrada do campo placa
+const placaMask = 'AAA-0#00'
+
+// Máscara para CPF: '000.000.000-00'
+// Máscara para CNPJ: '00.000.000/0000-00'
+
+export default function KarangosForm() {
+  const classes = useStyles()
+
+  const [karango, setKarango] = useState({
+    id: null,
+    marca: '',
+    modelo: '',
+    cor: '',
+    ano_fabricacao: (new Date()).getFullYear(), // Ano corrente
+    importado: '0',
+    placa: '',
+    preco: 0
+  })
+  const [currentId, setCurrentId] = useState()
+  const [importadoChecked, setImportadoChecked] = useState()
+
   const [snackState, setSnackState] = useState({
     open: false,
     severity: 'success',
-    message: 'Karango excluído com sucesso'  
+    message: 'Karango salvo com sucesso!'
   })
 
-  function handleDialogClose(result) {
-    setDialogOpen(false)
-    if(result) deleteItem()
-  }
+  const [btnSendState, setBtnSendState] = useState({
+    disabled: false,
+    label: 'Enviar'
+  })
 
-  function handleDeleteClick(id) {
-    setDeletable(id)
-    setDialogOpen(true)
-  }
+  const [error, setError] = useState({
+    marca: '',
+    modelo: '',
+    cor: '',
+    placa: '',
+    preco: ''
+  })
 
-  async function deleteItem() {
+  const [isModified, setIsModified] = useState(false)
+
+  const [dialogOpen, setDialogOpen] = useState(false) // O diálogo de confirmação está aberto?
+
+  const [title, setTitle] = useState('Cadastrar Novo Karango')
+
+  const history = useHistory()
+  const params = useParams()
+
+  useEffect(() => {
+    // Verifica se tem o parâmetro id na rota. Se tiver, temos que buscar
+    // os dados do registro no back-end para edição
+    if (params.id){
+      setTitle('Editando Karango')
+      getData(params.id)
+    } 
+  }, [])
+
+  async function getData(id) {
     try {
-      await axios.delete(`https://api.faustocintra.com.br/karangos/${deletable}`)
-      getData()   // Atualiza os dados da tabela
-      setSnackState({...snackState, open: true}) // Exibe a snackbar de sucesso    
+      let response = await axios.get(`https://api.faustocintra.com.br/karangos/${id}`)
+      setKarango(response.data)
     }
-    catch(error) {
-      // Mostra a snackbar de erro
+    catch (error) {
+      setSnackState({
+        open: true,
+        severity: 'error',
+        message: 'Não foi possível carregar os dados para edição.'
+      })
+    }
+  }
+
+  function handleInputChange(event, property) {
+
+    const karangoTemp = { ...karango }
+    let importadoCheckedTemp = importadoChecked
+
+    // Se houver id no event.target, ele será o nome da propriedade
+    // senão, usaremos o valor do segundo parâmetro
+    if (event.target.id) property = event.target.id
+
+    if (property === 'importado') {
+      const newState = !importadoChecked
+      //setKarango({ ...karango, importado: (newState ? '1' : '0') })
+      karangoTemp.importado = (newState ? '1' : '0')
+      //setImportadoChecked(newState)
+      importadoCheckedTemp = newState
+    }
+    else if (property === 'placa') {
+      //setKarango({ ...karango, [property]: event.target.value.toUpperCase() })
+      karangoTemp[property] = event.target.value.toUpperCase()
+    }
+    else {
+      // Quando o nome de uma propriedade de um objeto aparece entre [],
+      // isso se chama "propriedade calculada". O nome da propriedade vai
+      // corresponder à avaliação da expressão entre os colchetes
+      setCurrentId(event.target.id)
+      setKarango({ ...karango, [property]: event.target.value })
+      karangoTemp[property] = event.target.value
+    }
+    setKarango(karangoTemp)
+    setImportadoChecked(importadoCheckedTemp)
+    setIsModified(true) //o formulario foi modificado
+    validate(karangoTemp) // Dispara a validação
+  }
+
+  function validate(data) {
+
+    const errorTemp = {
+      marca: '',
+      modelo: '',
+      cor: '',
+      placa: '',
+      preco: ''
+    }
+    let isValid = true
+
+    // trim(): retira os espaços em branco do inicio e do final de uma string
+    // Validação do campo marca
+    if (data.marca.trim() === '') {
+      errorTemp.marca = 'A marca deve ser preenchida'
+      isValid = false
+    }
+
+    // Validação do campo modelo
+    if (data.modelo.trim() === '') {
+      errorTemp.modelo = 'O modelo deve ser preenchido'
+      isValid = false
+    }
+
+    // Validação do campo cor
+    if (data.cor.trim() === '') {
+      errorTemp.cor = 'Escolha uma cor'
+      isValid = false
+    }
+
+    // Validação do campo placa
+    // Valor válido não pode ser string vazia nem conter o caractere _
+    if (data.placa.trim() === '' || data.placa.includes('_')) {
+      errorTemp.placa = 'A placa deve ser corretamente preenchida'
+      isValid = false
+    }
+
+    // Validação do campo preco
+    // Valor válido deve ser numério e maior que zero
+    if (isNaN(data.preco) || Number(data.preco) <= 0) {
+      errorTemp.preco = 'O preço deve ser preenchido e maior que zero'
+      isValid = false
+    }
+    setError(errorTemp)
+    return isValid
+  }
+
+
+
+  function years() {
+    let result = []
+    for (let i = (new Date()).getFullYear(); i >= 1900; i--) result.push(i)
+    return result
+  }
+
+  async function saveData() {
+    try {
+      // Desabilitar o botão enviar
+      setBtnSendState({ disabled: true, label: 'Enviando...' })
+
+      // se o registro já existe (edição, verbo HTTP PUT)
+      if(params.id) await axios.put(`https://api.faustocintra.com.br/karangos/${params.id}`, karango)
+      // Registro não existe, cria um novo (verbo HTTP POST)
+      else await axios.post('https://api.faustocintra.com.br/karangos', karango)
+
+      setSnackState({
+        open: true,
+        severity: 'success',
+        message: 'Karango salvo com sucesso!'
+      })
+
+    }
+    catch (error) {
       setSnackState({
         open: true,
         severity: 'error',
         message: 'ERRO: ' + error.message
       })
     }
+    // Reabilitar o botão enviar
+    setBtnSendState({ disabled: false, label: 'Enviar' })
   }
+  function handleSubmit(event) {
 
-  async function getData() {
-    try {
-      let response = await axios.get('https://api.faustocintra.com.br/karangos?by=marca,modelo')
-      if(response.data.length > 0) setKarangos(response.data)
-    }
-    catch(error) {
-      console.error(error)
-    }
+    event.preventDefault() // Evitar o recarregamento da página
+
+    // Só salva os dados se eles forem válidos
+    if (validate(karango)) saveData()
   }
-
-  useEffect(() => {
-    getData()
-  }, [])  // Quando a dependência de um useEffect é um vetor vazio, isso indica
-          // que ele será executado apenas uma vez, na inicialização do componente
 
   function Alert(props) {
     return <MuiAlert elevation={6} variant="filled" {...props} />;
@@ -105,77 +254,152 @@ export default function KarangosList() {
 
   function handleSnackClose(event, reason) {
     // Evita que a snackbar seja fechada clicando-se fora dela
-    if(reason === 'clickaway') return
-    setSnackState({...snackState, open: false}) // Fecha a snackbar
+    if (reason === 'clickaway') return
+    setSnackState({ ...snackState, open: false }) // Fecha a snackbar
+    // Retorna à página de listagem
+    history.push('/list') // Retorna à página de listagem
   }
 
+  function handleDialogClose(result) {
+    setDialogOpen(false)
+
+    // Se o usuário concordou com a exclusão 
+    if (result) history.push('/list')
+  }
+
+  function handleGoBack() {
+
+    // Se o formulário estiver modificado, mostramos o diálogo de confirmação
+    if (isModified) setDialogOpen(true)
+    //senão, voltamos diretamente à página de listagem
+    else history.push('/list')
+  }
   return (
     <>
       <ConfirmDialog isOpen={dialogOpen} onClose={handleDialogClose}>
-        Deseja realmente excluir este karango?
+        Há dados não salvos. Deseja realmente voltar?
       </ConfirmDialog>
-      
-      <Snackbar open={snackState.open} autoHideDuration={6000} onClose={handleSnackClose}>
+
+      <Snackbar open={snackState.open} autoHideDuration={4000} onClose={handleSnackClose}>
         <Alert onClose={handleSnackClose} severity={snackState.severity}>
           {snackState.message}
         </Alert>
       </Snackbar>
-      
-      <h1>Listagem de Karangos</h1>
-      <Toolbar className={classes.toolbar} >
-        <Button color="secondary" variant="contained" size="large" 
-          startIcon={<AddBoxIcon />} onClick={() => history.push("/new") }>
-          Novo Karango
-        </Button>
-      </Toolbar>
-      <TableContainer component={Paper}>
-      <Table className={classes.table} size="small" aria-label="a dense table">
-        <TableHead>
-          <TableRow>
-            <TableCell align="right">Cód.</TableCell>
-            <TableCell>Marca</TableCell>
-            <TableCell>Modelo</TableCell>
-            <TableCell>Cor</TableCell>
-            <TableCell align="center">Ano</TableCell>
-            <TableCell align="center">Importado?</TableCell>
-            <TableCell align="center">Placa</TableCell>
-            <TableCell align="right">Preço</TableCell>
-            <TableCell align="center">Editar</TableCell>
-            <TableCell align="center">Excluir</TableCell>
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {
-            karangos.map(karango => 
-              <TableRow key={karango.id} className={classes.tableRow}>
-                <TableCell align="right">{karango.id}</TableCell>
-                <TableCell>{karango.marca}</TableCell>
-                <TableCell>{karango.modelo}</TableCell>
-                <TableCell>{karango.cor}</TableCell>
-                <TableCell align="center">{karango.ano_fabricacao}</TableCell>
-                <TableCell align="center">
-                  <Checkbox checked={karango.importado === '1'} readOnly />
-                </TableCell>
-                <TableCell align="center">{karango.placa}</TableCell>
-                <TableCell align="right">
-                  { Number(karango.preco).toLocaleString('pt-br', {style: 'currency', currency: 'BRL'}) }
-                </TableCell>
-                <TableCell align="center">
-                  <IconButton aria-label="edit">
-                    <EditIcon />
-                  </IconButton>
-                </TableCell>
-                <TableCell align="center">
-                  <IconButton aria-label="delete" onClick={() => handleDeleteClick(karango.id)}>
-                    <DeleteIcon color="error" />
-                  </IconButton>
-                </TableCell>    
-              </TableRow>
-            )
-          }
-        </TableBody>
-      </Table>
-    </TableContainer>
+
+      <h1>{title}</h1>
+      <form className={classes.form} onSubmit={handleSubmit}>
+
+        <TextField
+          id="marca"
+          label="Marca"
+          variant="filled"
+          value={karango.marca}
+          onChange={handleInputChange}
+          fullWidth
+          required
+          error={error.marca !== ''}
+          helperText={error.marca}
+        />
+
+        <TextField
+          id="modelo"
+          label="Modelo"
+          variant="filled"
+          value={karango.modelo}
+          onChange={handleInputChange}
+          fullWidth
+          required
+          error={error.modelo !== ''}
+          helperText={error.modelo}
+        />
+
+        <TextField
+          id="cor"
+          label="Cor"
+          variant="filled"
+          value={karango.cor}
+          onChange={event => handleInputChange(event, 'cor')}
+          fullWidth
+          select
+          required
+          error={error.cor !== ''}
+          helperText={error.cor}
+        >
+          <MenuItem value="Amarelo">Amarelo</MenuItem>
+          <MenuItem value="Azul">Azul</MenuItem>
+          <MenuItem value="Bege">Bege</MenuItem>
+          <MenuItem value="Branco">Branco</MenuItem>
+          <MenuItem value="Cinza">Cinza</MenuItem>
+          <MenuItem value="Dourado">Dourado</MenuItem>
+          <MenuItem value="Laranja">Laranja</MenuItem>
+          <MenuItem value="Marrom">Marrom</MenuItem>
+          <MenuItem value="Prata">Prata</MenuItem>
+          <MenuItem value="Preto">Preto</MenuItem>
+          <MenuItem value="Roxo">Roxo</MenuItem>
+          <MenuItem value="Verde">Verde</MenuItem>
+          <MenuItem value="Vermelho">Vermelho</MenuItem>
+        </TextField>
+
+        <TextField id="ano_fabricacao" label="Ano de Fabricacao" variant="filled" value={karango.ano_fabricacao} onChange={event => handleInputChange(event, 'ano_fabricacao')} fullWidth select>
+          {years().map(year => <MenuItem value={year} key={year}>{year}</MenuItem>)}
+        </TextField>
+
+        <TextField
+          id="preco"
+          label="Preço"
+          variant="filled"
+          value={karango.preco}
+          onChange={handleInputChange}
+          fullWidth
+          type="number"
+          onFocus={event => event.target.select()}
+          InputProps={{
+            startAdornment: <InputAdornment position="start">R$</InputAdornment>,
+          }}
+          required
+          error={error.preco !== ''}
+          helperText={error.preco}
+        />
+
+        <InputMask
+          formatChars={formatChars}
+          mask={placaMask}
+          id="placa"
+          onChange={event => handleInputChange(event, 'placa')}
+          value={karango.placa}
+        >
+          {() => <TextField
+            label="Placa"
+            variant="filled"
+            fullWidth
+            required
+            error={error.placa !== ''}
+            helperText={error.placa}
+          />}
+        </InputMask>
+
+        <FormControl className={classes.checkBox} fullWidth>
+          <FormControlLabel
+            control={<Checkbox checked={importadoChecked} onChange={handleInputChange} id="importado" />}
+            label="Importado?"
+          />
+        </FormControl>
+
+        <Toolbar className={classes.toolbar}>
+          <Button
+            variant="contained"
+            color="secondary"
+            type="submit"
+            disabled={btnSendState.disabled}
+          >
+            {btnSendState.label}</Button>
+          <Button variant="contained" onClick={handleGoBack}>
+            Voltar
+            </Button>
+        </Toolbar>
+
+        <div>{JSON.stringify(karango)}<br />currentId: {currentId}</div>
+      </form>
     </>
   )
 }
